Migrates RpcContext.logId from Integer to Long
This was discussed in #167.

+++ brpc-java-communication/src/main/java/com/baidu/brpc/RpcContext.java
@@ -48,7 +48,7 @@ public class RpcContext {
     /**
      * logId of protocol, application can set it.
      */
-    private Integer logId;
+    private Long logId;
 
     /**
      * set custom service instance tag,
